Improve specs
- Move to proper path
- Increase coverage
